test(mcp-server): split install.integration into 3 concern-scoped siblings (SMI-5263) (#1419)
install.integration.test.ts had grown to 948 lines (grandfathered in
scripts/check-file-length.ignore during the SMI-5261 mock-seam fix). Split along
the natural seam — the three module mocks (install.helpers, core
skill-installation.io, core install resolvers) and the only real installSkill()
usage were confined to the UUID install-path block:

- install.integration.test.ts (354) — filesystem/manifest/fetch-mocking
  primitives; no module mocks.
- install.parsing.integration.test.ts (166, new) — pure parseRepoUrl /
  parseSkillId logic.
- install.execution.integration.test.ts (464, new) — real installSkill() flow +
  trust-tier scanning; owns the three module mocks.

De-listed install.integration.test.ts from check-file-length.ignore (all three
now < 500). Suite unchanged: 174 passed / 6 skipped (12 files).
