+++ scripts/check-file-length.ignore
@@ -30,6 +30,6 @@
 # SMI-5211 split follow-up
 packages/core/tests/RawUrlSourceAdapter.security.test.ts
 
-# SMI-5263 split follow-up (pre-existing over-limit; grew during the SMI-5261
-# mock-seam fix — split into 3 concern-scoped siblings tracked in SMI-5263)
-packages/mcp-server/tests/integration/install.integration.test.ts
+# (SMI-5263 split follow-up cleared 2026-06-12: install.integration.test.ts was
+#  split into 3 concern-scoped siblings under 500 lines each —
+#  install.{parsing,execution}.integration.test.ts; suite still 174 passed / 6 skipped.)
